fix: Extract EntityOrmField as InstanceData.
