+++ packages/orm/src/index.ts
@@ -20,7 +20,7 @@ export { newPgConnectionConfig } from "joist-utils";
 export { AliasAssigner } from "./AliasAssigner";
 export * from "./Aliases";
 export { BaseEntity, getOrmField } from "./BaseEntity";
-export { Entity, EntityOrmField, IdType, isEntity } from "./Entity";
+export { Entity, IdType, isEntity } from "./Entity";
 export * from "./EntityFields";
 export * from "./EntityFilter";
 export * from "./EntityGraphQLFilter";
@@ -28,6 +28,7 @@ export * from "./EntityManager";
 export * from "./EntityMetadata";
 export { EnumMetadata } from "./EnumMetadata";
 export { EntityOrId, HintNode } from "./HintTree";
+export { EntityOrmField } from "./InstanceData";
 export * from "./QueryBuilder";
 export * from "./QueryParser";
 export * from "./changes";
